refactor: move whatsapp owner commands into platform
